requirements.txt
streamlit
python-dotenv
qdrant-client
sentence-transformers
langchain-community
PyMuPDF
google-generativeai==0.8.3
protobuf==4.25.3
cohere